Be cleverer in working out the correct length for ripping episodes.
A lot of DVDs have a "fake" track which is all the episodes strung
together. Great for viewing in a DVD but breaks when you want a digital
library. In stead of following the longest track heuristic we now calculate
the modal length (of tracks rounded to 5 minutes). And then use that as the
basic of our min/max time calculation.